optimize picture

+++ img/convert_to_webp.sh
@@ -1,6 +1,11 @@
 #!/bin/bash
-for file in ./*.jpg
+
+# Root img 
+for file in ./*.jpg ./*.png ./*/*/*.jpg ./*/*/*.png
 do
-  cp $file $file.bak
-  cwebp $file -o "${file%.*}.jpg"
+  if [ ! -f $file.bak ]
+  then
+    cp $file $file.bak
+    cwebp $file -o "$file"
+  fi
 done
